move src/read_to_end.rs -> src/io/read_to_end.rs
